+++ lib/simplecov.rb
@@ -52,6 +52,8 @@ module SimpleCov
     def start(profile = nil, &block)
       require "coverage"
       initial_setup(profile, &block)
+      require_relative "./simplecov/process" if SimpleCov.enabled_for_subprocesses?
+
       @result = nil
       self.pid = Process.pid