+++ test_projects/subprocesses/.simplecov_spawn.rb
@@ -1,0 +1,5 @@
+require 'bundler/setup'
+require 'simplecov'
+SimpleCov.command_name 'spawn'
+SimpleCov.at_fork.call(Process.pid)
+SimpleCov.start